Fix guild mark (x64) + guild symbol upload
Guild MARK ("upload sign") -- x64 regression: MarkImage.h typedef'd the pixel
as `unsigned long`, which is 8 bytes on the LP64 FreeBSD server but the mark
block format is fixed 4-byte BGRA (DevIL IL_BGRA/IL_UNSIGNED_BYTE, and the
12288-byte block on the wire). On LP64 every SGuildMarkBlock buffer was 2x too
large with an uninitialised upper half, so LZO compression + CRC ran over
garbage at the wrong size and the uploaded mark never displayed/updated.
Pin Pixel to uint32_t in both server and client MarkImage.h (4 bytes on every
data model; a no-op on the LLP64 client, the actual fix on the LP64 server).

Guild SYMBOL ("upload flag") -- the ".jpg not supported" failure: __LoadSymbol
called ilShutDown() after validating an image, tearing down the process-global
DevIL that is ilInit()'d once at startup; every subsequent ilLoad/GetImageInfo
then failed -> "The game does not support this picture." Remove the stray
ilShutDown(). Also fix a fall-through NULL FILE* deref on fopen_s failure
(missing early return) and a new BYTE[] / scalar-delete mismatch in the same
file. (lib/DevIL.dll was also replaced with the JPEG/PNG-capable build.)

+++ Lead-Client-Source/UserInterface/MarkImage.h
@@ -2,8 +2,12 @@
 #define __INC_METIN_II_MARKIMAGE_H__
 
 #include <IL/il.h>
+#include <cstdint>
 
-typedef unsigned long Pixel;
+// 32-bit BGRA pixel (match the server + the 4-byte-per-pixel mark/symbol wire format).
+// Was `unsigned long` -- 4 bytes on Win64/LLP64 so unchanged here, but pinned to a
+// fixed width to stay identical to the LP64 server.
+typedef uint32_t Pixel;
 
 struct SGuildMark
 {
